backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
idna==2.8
joblib==0.13.2
numpy==1.17.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.25.3
